pyproject.toml
[tool.poetry]
name = "hello-wordle"
version = "0.1.0"
description = "A wordle solver"
authors = ["Ben Strutt <ben.strutt1@nhs.net>"]

[tool.poetry.dependencies]
python = "^3.8"
colorama = "^0.4.4"
progressbar2 = "^4.0.0"
numpy = "^1.22.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
